use .sync modifier on description
