+++ edit/rio.cattle.io.externalservice.vue
@@ -72,10 +72,11 @@ export default {
     </div>
     <template v-else>
       <NameNsDescription
-        :value="value"
+        v-model="value.metadata"
         :mode="mode"
         name-label="Name"
         :register-before-hook="registerBeforeHook"
+        :description.sync="description"
       />
 
       <div class="spacer"></div>
