+++ edit/rio.cattle.io.stack.vue
@@ -75,10 +75,11 @@ export default {
     <div v-if="loading" />
     <form v-else>
       <NameNsDescription
-        :value="value"
+        v-model="value.metadata"
         :mode="mode"
         name-label="Stack Name"
         :register-before-hook="registerBeforeHook"
+        :description.sync="description"
       />
 
       <div class="spacer"></div>
